fix:Fix data type

+++ src/back-end/FBI/api/models.py
@@ -88,14 +88,14 @@ class Result(models.Model):
     viewedDate = models.DateTimeField()
     dataPath = models.CharField(max_length=100)
     emotion = models.CharField(max_length=10)
-    happiness = models.FloatField()
-    sadness = models.FloatField()
-    anger = models.FloatField()
-    contempt = models.FloatField()
-    disgust = models.FloatField()
-    fear = models.FloatField()
-    neutral = models.FloatField()
-    surprise = models.FloatField()
+    happiness = models.IntegerField()
+    sadness = models.IntegerField()
+    anger = models.IntegerField()
+    contempt = models.IntegerField()
+    disgust = models.IntegerField()
+    fear = models.IntegerField()
+    neutral = models.IntegerField()
+    surprise = models.IntegerField()
 
     class Meta:
         db_table = 'Result'
